Create end2end_tfkeras_serialize.py
When I read the Keras example I was initially unsure how to avoid the use of subprocess/saving the keras model to file. This is a minor modification to the existing example as I suspect some users would prefer not to go via disk, or may not have entitlements to do so.