Transpose optimization for Softmax and LogSoftmax (fixes #1716)
In opsets 13 and higher, the axis of the operation is arbitrary and can simply be changed according to the permutation of the Transpose.
In lower opsets, Softmax always coerces its inputs to a 2D tensor, making Transpose operations necessary if the permutation moves axes between the coerced batch and feature dimensions.

Signed-off-by: fthielke <fthielke@fb3.uni-bremen.de>