Test of negative axis concatenation where the first input has none shape.
In the test, a tensor with none shape is created by applying the slice
operator to another tensor. The slice size is obtained by adding two
tensors and therefore the size of the sliced tensor cannot be inferred and
is set to None. A tensor with a known shape is then concatenated to the
tensor with none shape along the -1 axis. Previously, it was not possible
to convert a network with this structure, but I fixed that bug in my
previous commit. The problem occurred only when the first input had none
shape and when the concatenation axis was -1.

Signed-off-by: Klas Magnusson <klamag@raysearchlabs.com>