Fix None shape error when one input to ConcatV2 has a shape. (#2135)
* Fix None shape error when one input to ConcatV2 has a shape.

I tried to convert a network where one of the inputs to a concatenation
along dimension -1 had shape None. The other input did however have a
shape. The conversion failed because the code only looked att the shape of
the first input to determine what positive axis value to use in the
concatenation. If the order of the inputs had been reversed, the conversion
would have worked.

I have now changed the code to look at the shapes of both input nodes. With
the new code, I can convert the network. I have also verified that the
resulting onnx-file works.

Signed-off-by: Klas Magnusson <klamag@raysearchlabs.com>
---------

Signed-off-by: Klas Magnusson <klamag@raysearchlabs.com>